+++ commands/forceStopMusic.js
@@ -1,29 +1,33 @@
 module.exports.run = async utils => {
 
-  let message = utils.message;
-  let bot = utils.bot;
-  let embeds = utils.embeds;
-  let serverQueue = utils.serverQueue;
+  if (utils.MySQL.cmdsEnabled.music) {
 
-  let voiceChannel = message.member.voiceChannel;
+    const message = utils.message;
+    const bot = utils.bot;
+    const embeds = utils.embeds;
+    const serverQueue = utils.serverQueue;
 
-  message.delete();
+    const voiceChannel = message.member.voiceChannel;
 
-  if (!voiceChannel) return message.channel.send(embeds.notinvoicechannelmusic(message.author.tag));
+    message.delete();
 
-  if (!voiceChannel.members.has(bot.user.id)) return message.channel.send(embeds.notinsamechannelmusic(message.author.tag));
+    if (!voiceChannel) return message.channel.send(embeds.notinvoicechannelmusic(message.author.tag));
 
-  if (!serverQueue || serverQueue.songs.length < 1) message.channel.send(embeds.nosongsqueuedmusic(message.author.tag));
+    if (!voiceChannel.members.has(bot.user.id)) return message.channel.send(embeds.notinsamechannelmusic(message.author.tag));
 
-  if (!message.member.hasPermission("MANAGE_MESSAGES")) return message.channel.send(embeds.noforcestoppermmusic(message.author.tag));
+    if (!serverQueue || serverQueue.songs.length < 1) message.channel.send(embeds.nosongsqueuedmusic(message.author.tag));
 
-  message.channel.send(embeds.successvotestop(message.author.tag, 1, 1));
+    if (!message.member.hasPermission("MANAGE_MESSAGES")) return message.channel.send(embeds.noforcestoppermmusic(message.author.tag));
 
-  serverQueue.songs = [];
+    message.channel.send(embeds.successvotestop(message.author.tag, 1, 1));
 
-  utils.updateQueue(serverQueue);
+    serverQueue.songs = [];
 
-  serverQueue.dispatcher.end("end");
+    utils.updateQueue(serverQueue);
+
+    serverQueue.dispatcher.end("end");
+
+  }
 
 };
